Transformed every int response to long

+++ src/main/kotlin/Day.kt
@@ -1,4 +1,4 @@
 abstract class Day {
-    abstract fun part1(): Int
-    abstract fun part2(): Int
+    abstract fun part1(): Long
+    abstract fun part2(): Long
 }
