+++ src/main/kotlin/day16/Day16.kt
@@ -3,11 +3,11 @@ package day16
 import Day
 
 class Day16(private val input: String) : Day() {
-    override fun part1(): Int {
+    override fun part1(): Long {
         return 0
     }
 
-    override fun part2(): Int {
+    override fun part2(): Long {
         return 0
     }
 }
