+++ src/main/kotlin/day25/Day25.kt
@@ -3,11 +3,11 @@ package day25
 import Day
 
 class Day25(private val input: String) : Day() {
-    override fun part1(): Int {
+    override fun part1(): Long {
         return 0
     }
 
-    override fun part2(): Int {
+    override fun part2(): Long {
         return 0
     }
 }
